+++ WebContent/cadastroUsuario.jsp
@@ -156,7 +156,7 @@
 							<c:out value="${user.nome}"></c:out>
 						</td>
 						<td>
-							<a href="salvarUsuario?acao=delete&user=${user.id}"><img src="resources/img/delete.png" 
+							<a href="salvarUsuario?acao=delete&user=${user.id}" onclick="return confirm('Confirma a exclusão?')"><img src="resources/img/delete.png" 
 							alt="Excluir" title="Excluir" width="20px" height="20px" /></a>
 						</td>
 						<td>
